Update Requirements.xlsx
Every crossing has exactly one building
Building types and sizes are not correlated
Building types and heights are uniformly distributed over the city
Only residents which are injured are created as agent instances in the model
The farthest building from the earthquake epicenter has a vulnerability of 0 due to the Building Vulnerability formula.
The amount of total residents injured and uninjured is saved in the model
The earthquake can be created on any place inside or outside of the city.