Ask all turtles which are part of network instead of only crossings
As ambulance, when selecting a new route, ask one-of all turtles that are part of the network, instead of only the crossings. This includes the hospitals, which fixes a bug when a ambulance drops of a patient and errors on requesting a new patient while (still) at the hospital.

Also use crossings-here instead of crossings-at 0 0. See http://ccl.northwestern.edu/netlogo/docs/dictionary.html#turtles-here