Record statistics on number and fraction of destroyed streets spotted
The number of roads spotted is significantly influenced by drones, and since drones are the main subject of our research and experiments, it's nice to track a direct metric (KPI) for the effect of drones. The indirect effect of course is the number of extra patients saved.

Also adds monitors and a histogram of it.